Show delete all chats menu entry when there is actually something
